dev/test_by_agent/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
livekit-api>=1.0.0
python-dotenv>=1.0.0
